+++ packages/core/avatar/src/components/__tests__/visual-regression/avatarSnapshotTest.js
@@ -5,7 +5,7 @@ import {
 } from '@atlaskit/visual-regression/helper';
 
 describe('Snapshot Test', () => {
-  it('basicAvatar-should match production example', async () => {
+  it('basicAvatar should match production example', async () => {
     const url = getExampleUrl(
       'core',
       'avatar',