+++ packages/core/inline-message/src/components/__tests__/visual-regression/messageSnapshotTest.js
@@ -1,0 +1,19 @@
+// @flow
+import {
+  getExampleUrl,
+  takeScreenShot,
+} from '@atlaskit/visual-regression/helper';
+
+describe('Snapshot Test', () => {
+  it('Inline message basic should match production example', async () => {
+    const url = getExampleUrl(
+      'core',
+      'inline-message',
+      'basic',
+      global.__BASEURL__,
+    );
+    const image = await takeScreenShot(global.page, url);
+    //$FlowFixMe
+    expect(image).toMatchProdImageSnapshot();
+  });
+});