+++ packages/elements/reactions/src/i18n/pt_BR.ts
@@ -1,0 +1,6 @@
+// Portuguese (Brazil)
+export default {
+  'fabric.reactions.loading': 'Carregando...',
+  'fabric.reactions.more.emoji': 'Mais emojis',
+  'fabric.reactions.error.unexpected': 'Algo deu errado',
+};