+++ packages/elements/reactions/src/i18n/sv.ts
@@ -1,0 +1,6 @@
+// Swedish
+export default {
+  'fabric.reactions.loading': 'Läser in ...',
+  'fabric.reactions.more.emoji': 'Fler emoji',
+  'fabric.reactions.error.unexpected': 'Något gick fel',
+};