+++ packages/media/media-picker/src/popup/reducers/__tests__/unit/removeFileFromRecents.spec.tsx
@@ -60,7 +60,7 @@ const extraState: Partial<State> = {
   },
   selectedItems: [
     {
-      serviceName: 'some-service',
+      serviceName: 'recent_files',
       mimeType: 'some-type',
       id: 'some-id',
       upfrontId: Promise.resolve('some-id'),
@@ -70,7 +70,7 @@ const extraState: Partial<State> = {
       occurrenceKey: 'some-other-occurrence-key',
     },
     {
-      serviceName: 'some-other-service',
+      serviceName: 'upload',
       mimeType: 'some-other-type',
       id: 'other-id',
       upfrontId: Promise.resolve('other-id'),