+++ packages/media/media-test-helpers/src/mocks/routers/api-router.ts
@@ -8,8 +8,8 @@ import {
   RequestHandler,
   Database,
 } from 'kakapo';
-import * as uuid from 'uuid';
 import { TouchFileDescriptor } from '@atlaskit/media-store';
+import * as uuid from 'uuid/v4';
 
 import { mapDataUriToBlob } from '../../utils';
 import { mockDataUri } from '../database/mockData';
@@ -286,12 +286,15 @@ export function createApiRouter(): Router<DatabaseSchema> {
         id: descriptor.id,
         collectionName: descriptor.collection,
       });
-      return {
-        type: 'file',
-        id: descriptor.id,
-        collection: descriptor.collection,
-        details: record.data.details,
-      };
+      if (record) {
+        return {
+          type: 'file',
+          id: descriptor.id,
+          collection: descriptor.collection,
+          details: record.data.details,
+        };
+      }
+      return null;
     });
 
     if (records.length) {
@@ -308,7 +311,11 @@ export function createApiRouter(): Router<DatabaseSchema> {
   router.post('/file/copy/withToken', (request, database) => {
     const { body, query } = request;
     const { sourceFile } = JSON.parse(body);
-    const { collection: destinationCollection } = query;
+    const {
+      collection: destinationCollection,
+      replaceFileId = uuid(),
+      occurrenceKey = uuid(),
+    } = query;
 
     const sourceRecord = database.findOne('collectionItem', {
       id: sourceFile.id,
@@ -318,9 +325,9 @@ export function createApiRouter(): Router<DatabaseSchema> {
     const { details, type, blob } = sourceRecord.data;
 
     const record = database.push('collectionItem', {
-      id: uuid.v4(),
+      id: replaceFileId,
       insertedAt: Date.now(),
-      occurrenceKey: uuid.v4(),
+      occurrenceKey,
       type,
       details,
       blob,