+++ packages/media/media-viewer/src/newgen/viewers/base-viewer.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { messages } from '@atlaskit/media-ui';
 import * as deepEqual from 'deep-equal';
-import { Context, ProcessedFileState } from '@atlaskit/media-core';
+import { Context, FileState } from '@atlaskit/media-core';
 import { Outcome } from '../domain';
 import ErrorMessage, { MediaViewerError } from '../error';
 import { Spinner } from '../loading';
@@ -10,7 +10,7 @@ import { ErrorViewDownloadButton } from '../download';
 
 export type BaseProps = {
   context: Context;
-  item: ProcessedFileState;
+  item: FileState;
   collectionName?: string;
 };