+++ packages/editor/renderer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/renderer
 
+## 33.0.3
+- Updated dependencies [00c648e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/00c648e):
+- Updated dependencies [a17bb0e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a17bb0e):
+- Updated dependencies [99f08a0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/99f08a0):
+  - @atlaskit/status@0.3.0
+
+## 33.0.2
+- [patch] [40510b0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/40510b0):
+
+  - Add panel type to fix copy-paste
+
 ## 33.0.1
 - Updated dependencies [135ed00](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/135ed00):
   - @atlaskit/editor-common@25.0.3